Example of Pickup item
This was just a class that I made for a game that I am working on right now that is a generic item that can be picked up by the player. The item takes in information from a data table that then updates the mesh and any other information needed to be updated immediately within the editor.